requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.37.3
botocore==1.37.3
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-filter==25.1
django-prometheus==2.3.1
django-redis==5.4.0
django-storages==1.14.5
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg==3.2.5
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
